+++ apps/esketch/frontend/XDHTML/core.cpp
@@ -55,7 +55,7 @@ namespace {
 		OnNotConnectedAllowedActions.Add(
 			xdhcmn::CloseActionLabel,
 			global::OnNewSession, global::About, global::Test,
-			prolog::DisplayProjectFilename, prolog::LoadProject, prolog::SwitchProjectType,	// All 'prolog'-related actions are allowed.
+			prolog::DisplayProjectFilename, prolog::LoadPreset, prolog::SwitchPreset,	// All 'prolog'-related actions are allowed.
 			login::Dismiss, login::DisplayEmbeddedBackendFilename, login::Connect, login::SwitchBackendType,	// All 'login'-related actions too.
 			main::HideTestButton, main::ShowTestButton, main::UnconnectedTesting );
 	};