+++ stable/sclm.h
@@ -229,18 +229,17 @@ namespace sclm {
 
 	void EraseProjectRegistry( void );
 
-	qENUM( ProjectType ) {
-		ptNew,			// Empty project.
-		ptPredefined,	// Use of a project defined in the 'Definitions' section in the configuration file.
-		ptRemote,		// Project stored in a file.
-		ptEmbedded,		// The content of the project is embedded in the its declaration (not implemented yet)
-		pt_amount,
-		pt_Undefined
+	qENUM( Preset ) {
+		pNone,			// No presets.
+		pSetup,		// Use a setup.
+		pProject,	// Use of a project file.
+		p_amount,
+		p_Undefined
 	};
 
-	const char *GetLabel( eProjectType ProjectType );
+	const char *GetLabel( ePreset Preset );
 
-	eProjectType GetProjectType( const str::string_ &Pattern );
+	ePreset GetPreset( const str::string_ &Pattern );
 
 	void LoadProject(
 		flw::iflow__ &Flow,
@@ -253,9 +252,9 @@ namespace sclm {
 		const scli::sInfo &Info,
 		str::string_ &Id );
 
-	void LoadProject(
-		eProjectType ProjectType,
-		const str::string_ &ProjectFeature,
+	void LoadPreset(
+		ePreset Preset,
+		const str::string_ &PresetFeature,
 		const scli::sInfo &Info );
 
 	void LoadProject( const scli::sInfo &Info );	// Load project, if applies, following configuration file indications.